refactor: authComponents
