+++ app/signup/page.tsx
@@ -1,15 +1,15 @@
 'use client';
 
-import AuthFooter from '@/components/auth/AuthFooter';
-import AuthHeader from '@/components/auth/AuthHeader';
+import Footer from '@/components/auth/Footer';
+import Header from '@/components/auth/Header';
 import RegisterForm from '@/components/auth/RegisterForm';
 
 export default function Login() {
   return (
     <>
-      <AuthHeader msg="첫 방문을 환영합니다!" />
+      <Header msg="첫 방문을 환영합니다!" />
       <RegisterForm />
-      <AuthFooter type="signup" />
+      <Footer type="signup" />
     </>
   );
 }
