+++ components/auth/Input.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
-import { PiEyeThin, PiEyeSlashThin } from 'react-icons/pi';
+import { IoMdEye, IoMdEyeOff } from 'react-icons/io';
 
 interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
-  usage: 'email' | 'password' | 'checkbox';
+  usage: 'email' | 'nickName' | 'password' | 'checkbox';
   hasError?: boolean;
 }
 
@@ -21,18 +21,18 @@ export default function Input({ usage, hasError = false, ...rest }: InputProps) 
     setPasswordType(passwordType === 'password' ? 'text' : 'password');
   };
 
-  if (usage === 'email') {
+  if (usage === 'email' || usage === 'nickName') {
     return <input className={signClassName(hasError)} {...rest} />;
   }
 
   if (usage === 'password') {
     return (
       <div className="relative w-[520px] h-[50px] max-w-full max-h-full">
-        <input className={signClassName(hasError)} {...rest} />
+        <input className={signClassName(hasError)} type={passwordType} {...rest} />
         {passwordType === 'password' ? (
-          <PiEyeSlashThin className={`${visibleButton} w-[24px] h-[24px]`} onClick={togglePasswordType} />
+          <IoMdEyeOff className={`${visibleButton} w-[24px] h-[24px]`} onClick={togglePasswordType} />
         ) : (
-          <PiEyeThin className={`${visibleButton} w-[24px] h-[24px]`} onClick={togglePasswordType} />
+          <IoMdEye className={`${visibleButton} w-[24px] h-[24px]`} onClick={togglePasswordType} />
         )}
       </div>
     );
